+++ rust/src/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod nnnoiseless;
+pub mod session;